# Vectorize `adjacent_find`