--- a/benchmarks/src/adjacent_find.cpp
+++ b/benchmarks/src/adjacent_find.cpp
@@ -7,7 +7,9 @@
 
 using namespace std;
 
-template <class T>
+enum class AlgType { Std, Rng };
+
+template <AlgType Alg, class T>
 void bm(benchmark::State& state) {
     const size_t size = static_cast<size_t>(state.range(0));
     const size_t pos  = static_cast<size_t>(state.range(1));
@@ -26,7 +28,11 @@ void bm(benchmark::State& state) {
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(v);
-        benchmark::DoNotOptimize(adjacent_find(v.begin(), v.end()));
+        if constexpr (Alg == AlgType::Std) {
+            benchmark::DoNotOptimize(adjacent_find(v.begin(), v.end()));
+        } else {
+            benchmark::DoNotOptimize(ranges::adjacent_find(v));
+        }
     }
 }
 
@@ -34,9 +40,14 @@ void common_args(auto bm) {
     bm->ArgPair(2525, 1142);
 }
 
-BENCHMARK(bm<char>)->Apply(common_args);
-BENCHMARK(bm<short>)->Apply(common_args);
-BENCHMARK(bm<int>)->Apply(common_args);
-BENCHMARK(bm<long long>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Std, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Std, short>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Std, int>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Std, long long>)->Apply(common_args);
+
+BENCHMARK(bm<AlgType::Rng, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Rng, short>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Rng, int>)->Apply(common_args);
+BENCHMARK(bm<AlgType::Rng, long long>)->Apply(common_args);
 
 BENCHMARK_MAIN();