--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -544,22 +544,26 @@ _NODISCARD _CONSTEXPR20 _FwdIt adjacent_find(const _FwdIt _First, _FwdIt _Last, 
     if (_UFirst != _ULast) {
 #if _USE_STD_VECTOR_ALGORITHMS
         if constexpr (_Equal_memcmp_is_safe<_FwdIt, _FwdIt, _Pr>) {
-            const auto _First_ptr = _STD _To_address(_UFirst);
-            const auto _Result    = _STD _Adjacent_find_vectorized(_First_ptr, _STD _To_address(_ULast));
+            if (!_STD _Is_constant_evaluated()) {
+                const auto _First_ptr = _STD _To_address(_UFirst);
+                const auto _Result    = _STD _Adjacent_find_vectorized(_First_ptr, _STD _To_address(_ULast));
 
-            if constexpr (is_pointer_v<decltype(_ULast)>) {
-                _ULast = _Result;
-            } else {
-                _ULast = _UFirst + (_Result - _First_ptr);
-            }
-        } else
-#endif // _USE_STD_VECTOR_ALGORITHMS
-        {
-            for (auto _UNext = _UFirst; ++_UNext != _ULast; _UFirst = _UNext) {
-                if (_Pred(*_UFirst, *_UNext)) {
-                    _ULast = _UFirst;
-                    break;
+                if constexpr (is_pointer_v<decltype(_ULast)>) {
+                    _ULast = _Result;
+                } else {
+                    _ULast = _UFirst + (_Result - _First_ptr);
                 }
+
+                _STD _Seek_wrapped(_Last, _ULast);
+                return _Last;
+            }
+        }
+#endif // _USE_STD_VECTOR_ALGORITHMS
+
+        for (auto _UNext = _UFirst; ++_UNext != _ULast; _UFirst = _UNext) {
+            if (_Pred(*_UFirst, *_UNext)) {
+                _ULast = _UFirst;
+                break;
             }
         }
     }