--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -6806,27 +6806,28 @@ namespace ranges {
 #if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (_Equal_memcmp_is_safe<_It, _It, _Pr> && sized_sentinel_for<_Se, _It>
                           && is_same_v<_Pj, identity>) {
-                const auto _First_ptr = _STD _To_address(_First);
-                const auto _Last_ptr  = _First_ptr + (_Last - _First);
+                if (!_STD is_constant_evaluated()) {
+                    const auto _First_ptr = _STD _To_address(_First);
+                    const auto _Last_ptr  = _First_ptr + (_Last - _First);
 
-                const auto _Result = _STD _Adjacent_find_vectorized(_First_ptr, _Last_ptr);
-                if constexpr (is_pointer_v<_It>) {
-                    return _Result;
-                } else {
-                    return _First + (_Result - _First_ptr);
+                    const auto _Result = _STD _Adjacent_find_vectorized(_First_ptr, _Last_ptr);
+                    if constexpr (is_pointer_v<_It>) {
+                        return _Result;
+                    } else {
+                        return _First + (_Result - _First_ptr);
+                    }
                 }
-            } else
+            }
 #endif // _USE_STD_VECTOR_ALGORITHMS
-            {
-                for (auto _Next = _First;; ++_First) {
 
-                    if (++_Next == _Last) {
-                        return _Next;
-                    }
+            for (auto _Next = _First;; ++_First) {
 
-                    if (_STD invoke(_Pred, _STD invoke(_Proj, *_First), _STD invoke(_Proj, *_Next))) {
-                        return _First;
-                    }
+                if (++_Next == _Last) {
+                    return _Next;
+                }
+
+                if (_STD invoke(_Pred, _STD invoke(_Proj, *_First), _STD invoke(_Proj, *_Next))) {
+                    return _First;
                 }
             }
         }