@StephanTLavavej 2025-03-17
Thanks! :smile_cat: I pushed minor changes and one performance fix to activate the optimization properly in C++14/17 mode.